Updated Survey Screen Corrected the survey screen, so it actually displays now Refactored everything so that the style is in the Style.js or MyComponent.js
