+++ client/package.json
@@ -21,6 +21,7 @@
     "react-native-multiple-select": "^0.5.12",
     "react-native-safe-area-context": "4.8.2",
     "react-native-screens": "~3.22.0",
+    "react-native-sectioned-multi-select": "^0.10.0",
     "react-native-splash-screen": "^3.3.0",
     "react-native-web": "~0.19.6",
     "typescript": "^5.1.3"
