+++ client/ui/screens/Randomizer.js
@@ -7,7 +7,7 @@ import { styles, colors } from '../assets/Style';
 export default function Randomizer({ navigation }) {
   return (
     <View style={{ flex: 1, backgroundColor: colors.black}}>
-		<Text>This is the Randomizer Screen</Text>
+		<Text style={styles.p}>This is the Randomizer Screen</Text>
     </View>
   );
 }
