+++ package.json
@@ -1,0 +1,6 @@
+{
+  "dependencies": {
+    "react-native-element-dropdown": "^2.10.2",
+    "react-native-sectioned-multi-select": "^0.10.0"
+  }
+}
